+++ groups/bsl/bslstl/bslstl_vector.t.cpp
@@ -6905,7 +6905,7 @@ void TestDriver<TYPE,ALLOC>::testCase9()
             int uOldLen = -1;
             for (int ui = 0; SPECS[ui]; ++ui) {
                 const char *const U_SPEC = SPECS[ui];
-                const size_t    uLen   = (int) strlen(U_SPEC);
+                const size_t      uLen   = strlen(U_SPEC);
 
                 if (verbose) {
                     printf("\tFor lhs objects of length " ZU ":\t", uLen);
@@ -6921,7 +6921,7 @@ void TestDriver<TYPE,ALLOC>::testCase9()
                 // int vOldLen = -1;
                 for (int vi = 0; SPECS[vi]; ++vi) {
                     const char *const V_SPEC = SPECS[vi];
-                    const size_t    vLen   = (int) strlen(V_SPEC);
+                    const size_t      vLen   = strlen(V_SPEC);
 
                     if (veryVerbose) {
                         printf("\t\tFor rhs objects of length " ZU ":\t",
@@ -7216,7 +7216,7 @@ void TestDriver<TYPE,ALLOC>::testCase7()
         int oldLen = -1;
         for (int ti = 0; SPECS[ti]; ++ti) {
             const char *const SPEC   = SPECS[ti];
-            const size_t    LENGTH = (int) strlen(SPEC);
+            const size_t      LENGTH = strlen(SPEC);
 
             if (verbose) {
                 printf("\nFor an object of length " ZU ":\n", LENGTH);
@@ -8156,10 +8156,10 @@ void TestDriver<TYPE,ALLOC>::testCase3()
 
         int oldLen = -1;
         for (int ti = 0; ti < NUM_DATA ; ++ti) {
-            const int          LINE   = DATA[ti].d_lineNum;
-            const char *const SPEC    = DATA[ti].d_spec_p;
-            const int         INDEX   = DATA[ti].d_index;
-            const size_t      LENGTH  = (int)strlen(SPEC);
+            const int          LINE  = DATA[ti].d_lineNum;
+            const char *const SPEC   = DATA[ti].d_spec_p;
+            const int         INDEX  = DATA[ti].d_index;
+            const size_t      LENGTH = strlen(SPEC);
 
             Obj mX(&testAllocator);
